README update, example image, shell fix
Updated the README to add a proper preamble, as well as usage and
building information.

Added an example image generated by me to the repository.

Updated the shell help message to also contain info about the cat and
stat commands. Also, a "man" command should be considered, to provide
more detailed information.